oauth2: fix passthrough logic to avoid modifying request when skipping filter (#40718)
### Description

This PR fixes a bug introduced in PR #40228, where OAuth2 cookies were
removed for requests matching the `pass_through_matcher` configuration.
This broke setups with multiple OAuth2 filter instances using different
`pass_through_matcher` configurations, because the first matching
instance removed the OAuth2 cookies--even when a passthrough was
intended--impacting subsequent filters that still needed those cookies.

The changes in this PR realign the filter behavior with the
documentation, which states that `pass_through_matcher` provides an
interface for users to specify header matching criteria such that, when
applicable, the OAuth flow is entirely skipped. When this occurs, only
the `oauth_passthrough` metric is incremented; no cookies or request
headers are modified, and no OAuth success is recorded.

In summary, when the `pass_through_matcher` configuration matches, the
filter should simply skip processing and leave the request untouched.

---

**Commit Message:** oauth2: fix passthrough logic to avoid modifying
request when skipping filter
**Risk Level:** Low
**Testing:** Unit test added and single page app test fixed by
[envoyproxy/examples PR
#819](https://github.com/envoyproxy/examples/pull/819)
**Docs Changes:** N/A
**Release Notes:** Added
**Platform Specific Features:** N/A

---------

Signed-off-by: Gustavo Moyano <gustavo.g.moyano@gmail.com>