Use gmock mock functions for testing callbacks (#40813)
Commit Message: Use gmock mock functions for testing callbacks
Additional Description: This is a testing-the-waters PR, is this change
desirable? If so I'll do the same to the rest of the file. The main hope
is that a good example existing would propagate next time someone
cargo-cults "how do I test this kind of thing".
(This change also slightly improves the test behavior in that now it
only passes if the initialize callback is called during `initialize`,
where before the test would also have passed if the initialize callback
was called during `dns_callback_`. This distinction seems like it might
be important based on the name of the test case and the comment
`Initialized without completing DNS resolution.`.)
Risk Level: test-only
Testing: test-only
Docs Changes: n/a
Release Notes: n/a
Platform Specific Features: n/a

---------

Signed-off-by: Raven Black <ravenblack@dropbox.com>